More stuff

+++ src/org/escaperun/game/model/items/AgilityBooster.java
@@ -14,7 +14,7 @@ public class AgilityBooster extends OneShotItem {
     @Override
     public void doAction(Entity e){
         Double amount = e.getStatContainer().getAgility().getBase() + 2.0;
-        e.getStatContainer().getAgility().setBase(amount);
+        e.getStatContainer().getAgility().setBase(amount.intValue());
     }
 
     @Override
